requirements.txt
fastapi
nicegui
uvicorn
beanie
motor
loguru
pydantic
python-dotenv
markdown
